fullstack-llm-google-sheets/requirements.txt
fastapi
uvicorn
transformers
torch
gspread
google-auth
pydantic
python-dotenv
